Bug fix for Storage/Location entities
Location doesn't use the '.' but the './' and every location SHOULD start with
a '/'.  To implement the Data Mesh governance we need to be able to search by
prefix not by service. Therefore the SQL query is not '.%' but just '%'.

While trying to implement the Operation library on top of OpenMetadata I have
encountered the Jersey client timeout. I have changed the
CatalogApplicationTest to have no timeout so that I can use the debugger. The
flow is Test -(http)> OperationAPI -(http)> OpenMetadataAPI -(reso/repo)> MySQL
and the timeout is between Test and OperationAPI. I've tried to modify the test
Dropwizard Configuration yaml file without success.